Shell collapse: hoist System Events counter-chart bodies to Ui.SystemHealthChartRenderer
The four ScottPlot render bodies behind the System Events tab's Corruption + Contention 2x2
chart grids (RenderCounterChart / RenderSickSpinlocksChart / RenderCpuComparisonChart /
AddNoDataText) were byte-identical between Lite ServerTab.SystemHealthCharts.cs and Darling
ViewerServerTab.SystemHealthCharts.cs apart from the display-time projection. Collapsed into
a shared PerformanceMonitor.Ui.SystemHealthChartRenderer (internal -- takes the internal
ChartHoverHelper, reached via Ui InternalsVisibleTo). The projection divergence (Lite's
.AddMinutes(UtcOffsetMinutes) vs Darling's ViewerTimeHelper.ForDisplay) is injected as a
Func<DateTime,DateTime>; the store-specific data read, the eight hover fields, and the
chart-field wiring stay per-app, each handing chart + hover + projection to the shared
renderer via a lazily-built SysHealthRenderer. Faithful extraction (exact prior statements,
only projection + chrome/legend helper calls rerouted). Net -156 lines across the two apps.

The deprecated Full Dashboard keeps its own copies.

Verified: all three apps build, both capability pins green (Lite 6/6, Darling 6/6). NOTE:
the pin proves the charts are wired but does not test rendered output, so a visual
spot-check of the Corruption/Contention sub-tabs is the belt-and-suspenders confirmation.
