+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **Front-end shell collapse: System Events counter-chart bodies hoisted to `Ui.SystemHealthChartRenderer`** ([#1527]) — the four ScottPlot render bodies behind the System Events tab's Corruption + Contention 2x2 chart grids (`RenderCounterChart` / `RenderSickSpinlocksChart` / `RenderCpuComparisonChart` / `AddNoDataText`), which were BYTE-IDENTICAL between Lite `ServerTab.SystemHealthCharts.cs` and Darling `ViewerServerTab.SystemHealthCharts.cs` apart from the display-time projection, collapse into a shared `PerformanceMonitor.Ui.SystemHealthChartRenderer` (~180 lines shared; net −156 across the two apps). The one real divergence — Lite's `.AddMinutes(UtcOffsetMinutes)` vs Darling's `ViewerTimeHelper.ForDisplay` — is injected as a `Func<DateTime, DateTime>`; the store-specific data read (`RefreshSystemHealthChartsAsync` / `LoadSystemHealthChartsAsync`), the eight `ChartHoverHelper` fields, and the chart-field wiring stay per-app, each handing its chart + hover + projection to the shared renderer via a lazily-built `SysHealthRenderer`. The renderer is `internal` (it takes the internal `ChartHoverHelper`; both apps reach it via Ui's `InternalsVisibleTo`). This is a FAITHFUL extraction — the shared bodies are the exact prior statements with only the projection and the chrome/legend helper calls rerouted (`ClearChart`/`ShowChartLegend` → the shared `ChartRenderHelper`, `ApplyTheme`/`ReapplyAxisColors`/`SetChartYLimitsWithLegendPadding` → the shared `ChartStyle`). The deprecated Full Dashboard keeps its own copies. Verified: all three apps build and the capability pin stays green in BOTH Lite (6/6) and Darling (6/6). NOTE: the pin proves the charts are WIRED but does not test rendered output (there is no chart-render unit test), so a visual spot-check of the Corruption / Contention sub-tabs is the belt-and-suspenders confirmation. Refactor only — no behavior change.
 - **Front-end shell collapse: stateful chart-render helpers hoisted to `Ui.ChartRenderHelper`** ([#1526]) — the ScottPlot legend-panel tracking (the `_legendPanels` dictionary + `ClearChart` / `ShowChartLegend`) and the empty-state "No Data" chart (`RefreshEmptyChart`), duplicated verbatim in Lite `ServerTab.Charts.cs` and Darling `ViewerServerTab.ChartHelpers.cs`, collapse into a shared `PerformanceMonitor.Ui.ChartRenderHelper`. Each app now holds one `ChartRenderHelper` instance and routes its (many) `ClearChart` / `RefreshEmptyChart` / `ShowChartLegend` call sites through thin per-app forwarders — call sites unchanged — while the Query Heatmap's direct legend-panel write goes through a new `SetLegendPanel` accessor. The stateless theme / axis-color / y-limit shapes stay as per-app forwarders (they already delegate to the shared `ChartStyle`). Net −116 lines. This is the prerequisite for sharing the per-tab chart BODIES (the System Events / system_health counter charts next), which call these helpers. The deprecated Full Dashboard keeps its own copies. Verified: all three apps build and the capability pin stays green in BOTH Lite (6/6) and Darling (6/6). Refactor only — no behavior change.
 - **Front-end shell collapse: file-save dialogs hoisted to `Ui.FileSaveHelper`** ([#1525]) — the byte-identical `SaveFileDialog` + `File.WriteAllText` + error-dialog blocks behind the plan / deadlock-XML / blocked-process-XML download buttons collapse into a shared `PerformanceMonitor.Ui.FileSaveHelper`. `SavePlanFile` (verbatim in Lite `ServerTab.Plans.cs` + Darling `ViewerServerTab.Plans.cs`) and the `DownloadDeadlockXml` / `DownloadBlockedProcessXml` handler bodies (four copies across the two apps) now go through `FileSaveHelper.SavePlanFile` / `SaveXmlToFile` via a per-file `using static`, so the `SavePlanFile(...)` call sites stay unchanged and each app keeps only its thin row-type-specific XML handler (extracting the per-app row's XML + building the file name). Net −114 lines. Behavior preserved exactly — same `.sqlplan` / `.xml` filters, `yyyyMMdd_HHmmss` naming, and error text. The deprecated Full Dashboard keeps its own inline copies (no `FileSaveHelper` reference, so no collision). Verified: all three apps build with 0 errors and the capability pin stays green in BOTH Lite (6/6) and Darling (6/6). Refactor only — no behavior change.
 - **Front-end shell collapse — Batch B: `FindParentDataGrid` visual-tree helpers hoisted to `PerformanceMonitor.Ui`** ([#1524]) — the second code de-duplication under the [#1522] capability pin. The two visual-tree-walk helpers that resolve a context-menu / element to its owning `DataGrid` — `FindParentDataGrid` (the owning grid of the clicked `MenuItem`) and `FindParentDataGridFromElement` (walk up from any element) — were copy-pasted TEN times across Lite and the Darling viewer (Lite: `ContextMenuHelper`, `AlertsHistoryTab`, `FinOpsTab` ×2, `ServerTab.CopyExport`, `ServerTab.Filters`; Darling: `AlertsHistoryTab`, `FinOpsTab`, `ViewerServerTab.CopyExport`, `ViewerServerTab.Filters`). Both now live ONCE in the shared `Ui` project as `DataGridHelpers`, consumed via a per-file `using static`. The shared `FindParentDataGrid` takes the superset `object` form (Lite's `ContextMenuHelper` signature) — behaviorally identical to the `MenuItem` copies for the MenuItem every caller passes — so all 19 caller files compile and behave unchanged. Because `ServerTab` / `ViewerServerTab` are PARTIAL classes and `using static` is per-file, every partial that CALLS a helper (including the block-chain / deadlock / plans / active-queries / live-plan / blocking partials that only call it, never defined it) got its own `using static`; the call sites were left untouched. The deprecated Full Dashboard has no copies and is untouched. Verified end-to-end: all three apps build with 0 errors and the capability pin stays green in BOTH Lite (6/6) and Darling (6/6). Refactor only — no behavior change.
@@ -323,6 +324,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1513]: https://github.com/erikdarlingdata/PerformanceMonitor/pull/1513
 [#1514]: https://github.com/erikdarlingdata/PerformanceMonitor/pull/1514
 [#1515]: https://github.com/erikdarlingdata/PerformanceMonitor/pull/1515
+[#1527]: https://github.com/erikdarlingdata/PerformanceMonitor/pull/1527
 [#1526]: https://github.com/erikdarlingdata/PerformanceMonitor/pull/1526
 [#1525]: https://github.com/erikdarlingdata/PerformanceMonitor/pull/1525
 [#1524]: https://github.com/erikdarlingdata/PerformanceMonitor/pull/1524
